fix(gemini): extract tool conversion to utils so unit tests skip API key
Unit tests for tool-schema conversion shared the integration `llm`
fixture, which constructs `GeminiLLM()` and requires `GOOGLE_API_KEY`.
CI sets the env var to empty, so setup failed. Move the pure conversion
function to `utils.py` and call it directly from the tests.
